scripts/performance_test/requirements.txt
GitPython
pandas
